% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{.simpleCap}
\alias{.simpleCap}
\title{Turn to uppercase the first letter of a string}
\usage{
.simpleCap(x)
}
\arguments{
\item{x:}{any string}
}
\value{
The string with first letter capital
}
\description{
Turn to uppercase the first letter of a string
}
\examples{
output <- .simpleCap ("the quick red fox jumps over the lazy brown dog");
}
