% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{drin_popp}
\alias{drin_popp}
\title{Create posterior predictions plot for Developmental Rate model}
\usage{
drin_popp(data, est, mtype = "bieri", lik = "gauss")
}
\arguments{
\item{rdevel}{object that contains (est) s4 type Stanfit object, (summary) Stanfit summary of basic parameters, (diagnostics) Stanfit overal diagnostics report}
}
\value{
Create posterior predictions plot for current model
}
\description{
Create posterior predictions plot for Developmental Rate model
}
\seealso{
\code{\link[drin_hmc]{stan}}
}
