% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{fanalytis_gauss}
\alias{fanalytis_gauss}
\title{#' A string that contains the R-Stan code for Analytis model 
with Gaussian likelihood}
\usage{
fanalytis_gauss()
}
\arguments{
\item{funcode:}{no parameter is required}
}
\value{
The output is the corresponding R-Stan code
}
\description{
#' A string that contains the R-Stan code for Analytis model 
with Gaussian likelihood
}
\examples{
output <- fanalytis_gauss();
}
