% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{fanalytis_ig}
\alias{fanalytis_ig}
\title{#' A string that contains the R-Stan code for Analytis model 
with Inverse Gamma likelihood. No zeros considered}
\usage{
fanalytis_ig()
}
\arguments{
\item{funcode:}{no parameter is required}
}
\value{
The output is the corresponding R-Stan code
}
\description{
#' A string that contains the R-Stan code for Analytis model 
with Inverse Gamma likelihood. No zeros considered
}
\examples{
output <- fanalytis_ig();
}
