% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{fanalytis_ziig}
\alias{fanalytis_ziig}
\title{#' A string that contains the R-Stan code for Analytis model 
with zero inflated Inverse Gamma likelihood. Excess zeros considered}
\usage{
fanalytis_ziig()
}
\arguments{
\item{funcode:}{no parameter is required}
}
\value{
The output is the corresponding R-Stan code
}
\description{
#' A string that contains the R-Stan code for Analytis model 
with zero inflated Inverse Gamma likelihood. Excess zeros considered
}
\examples{
output <- fanalytis_ziig();
}
