% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{fbieri_gauss}
\alias{fbieri_gauss}
\title{#' A string that contains the R-Stan code for Bieri model 
with Gaussian likelihood}
\usage{
fbieri_gauss()
}
\arguments{
\item{funcode:}{no parameter is required}
}
\value{
The output is the corresponding R-Stan code
}
\description{
#' A string that contains the R-Stan code for Bieri model 
with Gaussian likelihood
}
\examples{
output <- fbieri_gauss();
}
