% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{fbieri_ig}
\alias{fbieri_ig}
\title{#' A string that contains the R-Stan code for Bieri model 
with Inverse Gamma likelihood. No zeros considered}
\usage{
fbieri_ig()
}
\arguments{
\item{funcode:}{no parameter is required}
}
\value{
The output is the corresponding R-Stan code
}
\description{
#' A string that contains the R-Stan code for Bieri model 
with Inverse Gamma likelihood. No zeros considered
}
\examples{
output <- fbieri_ig();
}
