% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{fbriere_ig}
\alias{fbriere_ig}
\title{#' A string that contains the R-Stan code for Briere model 
with Inverse Gamma likelihood. No zeros considered}
\usage{
fbriere_ig()
}
\arguments{
\item{funcode:}{no parameter is required}
}
\value{
The output is the corresponding R-Stan code
}
\description{
#' A string that contains the R-Stan code for Briere model 
with Inverse Gamma likelihood. No zeros considered
}
\examples{
output <- fbriere_ig();
}
