% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drin_anth.R
\name{hush}
\alias{hush}
\title{An S4 class to represent a developmental rates object.
Supresses output of a function}
\usage{
hush(funcode)
}
\arguments{
\item{funcode:}{any function code}
}
\value{
The output of the function without displaying it
}
\description{
An S4 class to represent a developmental rates object.
Supresses output of a function
}
\examples{
output <- hush (print(1:100));
}
